pass timestamps as localtime to API
The smashrun API is a bit weird in that the timestamps that it
accepts in the fromDate query parameter are expected to be local time,
not UTC.

Now, the timestamp is just <current time> - <jan 1, 1970> and is
completely oblivious of time zones.

This fixes #6

Note that when they add fromDateUTC we should switch to that since it
makes more sense.