Remove client-side validation of style kwarg for get_activities
There was a type (brief vs briefs) in the style validation, which
highlighted the fact that the validation wasn't particularly helpful,
and worse just made things more fragile.

So i removed the validation. If the user passes in an invalid value for
summary, then the server will respond with a 404.

This fixes #9